refactor(automation): decompose ImplementationPhaseRunner into five phases
Re-implements the #712 decomposition on top of current main. The prior PR
branch was based on a 110+-commit-stale snapshot whose phase bodies predated
~20 correctness fixes to the review loop (GO-gate, model-in-session-key,
HUMAN_BLOCKED, secret-skip commit, raising push) and introduced two behavioral
regressions flagged in review. Rather than rebase the stale bodies, this extracts
main's CURRENT methods verbatim into single-responsibility collaborators so all
fixes are preserved by construction and both regressions vanish.

Splits the 2644-line god-class into:
- StageContext + StageMixin (_stage_context.py) — shared per-issue references
- PlanPhase (_plan_phase.py) — plan presence + generation
- ImplementPhase (_implement_phase.py) — advise + agent invocation
- ReviewPhase (_review_phase.py) — strict review + address loop, verdict labeling
- PRCreatePhase (_pr_create_phase.py) — pre-PR test gate + ensure-PR
- FollowUpPhase (_followup_phase.py) — /learn + follow-up filing

ImplementationPhaseRunner becomes a thin pipeline coordinator owning only the
top-level orchestration (_implement_issue / _review_existing_pr / dirty-worktree
salvage) and forwarding the original method names to the phases — preserving the
patch.object(impl, "_method") test-patch contract.

All `# noqa: C901` waivers in the touched code are removed (_implement_issue and
_run_impl_review_loop are broken into focused helpers). Adds isolated per-phase
unit tests (test_stage_phases.py) that exercise each phase against a lightweight
StageContext stub without mocking 30+ collaborators. Existing tests updated to
patch symbols at their new module homes.

Closes #712
