+++ hephaestus/automation/_followup_phase.py
@@ -1,0 +1,227 @@
+"""Post-PR follow-up phase (/learn + follow-up issue filing).
+
+Extracted from :class:`ImplementationPhaseRunner` as part of the #712
+decomposition. :class:`FollowUpPhase` owns everything that runs *after* the
+review loop converges: resuming the implementer session to run ``/learn``,
+compacting it, filing follow-up issues, and finally marking the issue
+``COMPLETED``. It also owns the cross-issue ``_rerun_failed_learns`` sweep.
+"""
+
+from __future__ import annotations
+
+import logging
+from datetime import datetime, timezone
+from pathlib import Path
+from typing import TYPE_CHECKING, Any
+
+from hephaestus.agents.runtime import is_codex, session_agent_matches
+
+from ._stage_context import StageMixin
+from .claude_models import implementer_model
+from .follow_up import parse_follow_up_items, run_follow_up_issues
+from .git_utils import issue_ref
+from .learn import learn_needs_rerun, run_learn
+from .models import ImplementationPhase, ImplementationState
+
+if TYPE_CHECKING:
+    from ._stage_context import StageContext
+
+logger = logging.getLogger(__name__)
+
+
+class FollowUpPhase(StageMixin):
+    """Run /learn + follow-up issue filing after a PR is created."""
+
+    def __init__(self, ctx: StageContext) -> None:
+        """Store the shared :class:`StageContext`."""
+        self.ctx = ctx
+
+    def _run_post_pr_followup(
+        self,
+        issue_number: int,
+        worktree_path: Path,
+        state: ImplementationState,
+        slot_id: int | None,
+    ) -> None:
+        """Run /learn and file follow-up issues after the PR is created."""
+        impl = self.impl
+        # Learn phase (after CREATING_PR, before COMPLETED)
+        can_resume_session = self.runner._can_resume_state_session(state)
+        if (
+            self.options.enable_learn
+            and not state.learn_completed
+            and can_resume_session
+            and state.session_id
+        ):
+            if slot_id is not None:
+                self.status_tracker.update_slot(
+                    slot_id, f"{issue_ref(issue_number)}: Running learn"
+                )
+            with self.state_lock:
+                state.phase = ImplementationPhase.LEARN
+            impl._save_state(state)
+            retro_success = self.runner._run_learn(
+                state.session_id,
+                worktree_path,
+                issue_number,
+                slot_id,
+                session_agent=state.session_agent,
+            )
+            with self.state_lock:
+                state.learn_completed = retro_success
+            impl._save_state(state)
+            if retro_success and not is_codex(self.options.agent):
+                self.runner._compact_implementer_session(issue_number, worktree_path)
+
+        # Follow-up issues phase (after LEARN, before COMPLETED)
+        if self.options.enable_follow_up and can_resume_session and state.session_id:
+            if slot_id is not None:
+                self.status_tracker.update_slot(
+                    slot_id, f"{issue_ref(issue_number)}: Identifying follow-ups"
+                )
+            with self.state_lock:
+                state.phase = ImplementationPhase.FOLLOW_UP_ISSUES
+            impl._save_state(state)
+            self.runner._run_follow_up_issues(
+                state.session_id,
+                worktree_path,
+                issue_number,
+                slot_id,
+                session_agent=state.session_agent,
+            )
+
+        # Mark as completed
+        with self.state_lock:
+            state.phase = ImplementationPhase.COMPLETED
+            state.completed_at = datetime.now(timezone.utc)
+        impl._save_state(state)
+
+    def _parse_follow_up_items(self, text: str) -> list[dict[str, Any]]:
+        """Parse follow-up items from Claude's JSON response."""
+        return parse_follow_up_items(text)
+
+    def _can_resume_state_session(self, state: ImplementationState) -> bool:
+        """Return True when the saved session can be resumed by the selected agent."""
+        if not state.session_id:
+            return False
+        if session_agent_matches(state.session_agent, self.options.agent):
+            return True
+        logger.info(
+            "Skipping session resume for issue #%s: session belongs to %s, selected agent is %s",
+            state.issue_number,
+            state.session_agent or "claude",
+            self.options.agent,
+        )
+        return False
+
+    def _run_follow_up_issues(
+        self,
+        session_id: str,
+        worktree_path: Path,
+        issue_number: int,
+        slot_id: int | None = None,
+        *,
+        session_agent: str | None = None,
+    ) -> None:
+        """Resume the selected agent session to identify and file follow-up issues."""
+        run_follow_up_issues(
+            session_id,
+            worktree_path,
+            issue_number,
+            self.state_dir,
+            self.status_tracker,
+            slot_id,
+            dry_run=self.options.dry_run,
+            agent=self.options.agent,
+            session_agent=session_agent,
+        )
+
+    def _learn_needs_rerun(self, issue_number: int) -> bool:
+        """Check if learn log indicates failure."""
+        return learn_needs_rerun(issue_number, self.state_dir)
+
+    def _rerun_failed_learns(self) -> dict[int, bool]:
+        """Re-run failed learns for completed issues.
+
+        Returns:
+            Dictionary mapping issue number to success status
+
+        """
+        impl = self.impl
+        results: dict[int, bool] = {}
+
+        for issue_number, state in self.impl.state_mgr.states.items():
+            # Only re-run for completed issues with failed learns
+            if (
+                state.phase != ImplementationPhase.COMPLETED
+                or state.learn_completed
+                or not self._can_resume_state_session(state)
+            ):
+                continue
+
+            # Check if log indicates failure
+            if not impl._learn_needs_rerun(issue_number):
+                continue
+
+            # Verify worktree exists
+            if not state.worktree_path:
+                logger.warning("Skipping learn re-run for #%s: no worktree_path", issue_number)
+                continue
+
+            session_id = state.session_id
+            if session_id is None:
+                continue
+
+            worktree_path = Path(state.worktree_path)
+            if not worktree_path.exists():
+                logger.warning("Skipping learn re-run for #%s: worktree not found", issue_number)
+                continue
+
+            # Re-run learn
+            logger.info("Re-running failed learn for issue #%s", issue_number)
+            success = impl._run_learn(
+                session_id,
+                worktree_path,
+                issue_number,
+                slot_id=None,
+                session_agent=state.session_agent,
+            )
+
+            # Update and save state
+            with self.state_lock:
+                state.learn_completed = success
+            impl._save_state(state)
+
+            results[issue_number] = success
+
+        if results:
+            success_count = sum(1 for s in results.values() if s)
+            logger.info(
+                "Re-ran %s learn(s): %s succeeded, %s failed",
+                len(results),
+                success_count,
+                len(results) - success_count,
+            )
+
+        return results
+
+    def _run_learn(
+        self,
+        session_id: str,
+        worktree_path: Path,
+        issue_number: int,
+        slot_id: int | None = None,
+        *,
+        session_agent: str | None = None,
+    ) -> bool:
+        """Resume the selected agent session to run /learn."""
+        return run_learn(
+            session_id,
+            worktree_path,
+            issue_number,
+            self.state_dir,
+            slot_id,
+            agent=self.options.agent,
+            session_agent=session_agent,
+            model=implementer_model(),
+        )
