+++ hephaestus/automation/_plan_phase.py
@@ -1,0 +1,98 @@
+"""Plan presence + generation phase for the implementation pipeline.
+
+Extracted from :class:`ImplementationPhaseRunner` as part of the #712
+decomposition. :class:`PlanPhase` owns the "does this issue already have an
+implementation plan, and if not, generate one" responsibility.
+"""
+
+from __future__ import annotations
+
+import contextlib
+import json
+import subprocess
+import sys
+from typing import TYPE_CHECKING
+
+from ._stage_context import StageMixin
+from .git_utils import run
+from .planner_state import _comments_contain_plan
+
+if TYPE_CHECKING:
+    from ._stage_context import StageContext
+
+
+class PlanPhase(StageMixin):
+    """Ensure an issue has an implementation plan before implementation."""
+
+    def __init__(self, ctx: StageContext) -> None:
+        """Store the shared :class:`StageContext`."""
+        self.ctx = ctx
+
+    def _has_plan(self, issue_number: int) -> bool:
+        """Check if issue has an implementation plan.
+
+        Delegates to :func:`planner_state._comments_contain_plan` so the
+        prefix-anchored check stays in sync with the planner. Substring
+        matching here previously caused the implementer to mistake a
+        ``## 🔍 Plan Review`` comment (which quotes the plan body) for the
+        plan itself — the same bug class fixed in #455/#468/#484 (#715).
+
+        Note: ``_comments_contain_plan`` is a private helper but is the
+        canonical implementation per its own docstring; cross-module reuse
+        here is intentional to avoid a third copy of the same prefix logic.
+        """
+        try:
+            result = run(
+                ["gh", "issue", "view", str(issue_number), "--comments", "--json", "comments"],
+                capture_output=True,
+            )
+            data = json.loads(result.stdout)
+            comments = data.get("comments", [])
+            return _comments_contain_plan(comments)
+        except (subprocess.SubprocessError, json.JSONDecodeError, OSError):
+            return False
+
+    def _generate(self, issue_number: int) -> None:
+        """Generate plan for an issue using hephaestus-plan-issues."""
+        import shutil
+
+        # Prefer the installed entry point (works in any repo)
+        entry_point = shutil.which("hephaestus-plan-issues")
+        if entry_point:
+            run(
+                [entry_point, "--issues", str(issue_number), "--agent", self.options.agent],
+                timeout=600,
+            )
+            return
+
+        # Fall back to python -m invocation (works when PYTHONPATH is set).
+        # On failure, fall through to the legacy scripts/plan_issues.py path.
+        with contextlib.suppress(subprocess.SubprocessError, OSError):
+            run(
+                [
+                    sys.executable,
+                    "-m",
+                    "hephaestus.automation.planner",
+                    "--issues",
+                    str(issue_number),
+                    "--agent",
+                    self.options.agent,
+                ],
+                timeout=600,
+            )
+            return
+
+        # Legacy fallback: local scripts/plan_issues.py (ProjectScylla layout)
+        plan_script = self.repo_root / "scripts" / "plan_issues.py"
+        if plan_script.exists():
+            run(
+                [sys.executable, str(plan_script), "--issues", str(issue_number)],
+                timeout=600,
+            )
+            return
+
+        raise RuntimeError(
+            "Could not find hephaestus-plan-issues entry point, "
+            "hephaestus.automation.planner module, or "
+            f"scripts/plan_issues.py in {self.repo_root}"
+        )
